+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [0.4.0]
+
 ### Added
 * IronPython output is now redirected to the Unity console
 
@@ -46,7 +49,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial support for Python in Unity3D with IronPython 2.6.2.
 
-[Unreleased]: https://github.com/exodrifter/unity-python/compare/0.3.1...HEAD
+[Unreleased]: https://github.com/exodrifter/unity-python/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/exodrifter/unity-python/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/exodrifter/unity-python/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/exodrifter/unity-python/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/exodrifter/unity-python/compare/0.1.0...0.2.0